Allow reading image over serial connection
These changes are only effective if the microsd card is removed. Instead of displaying the test image, the controller listens on the serial connection with a timeout of 1 second.

I don't quite understand why the "wakeup byte" doesn't shift the image but it doesn't seem to.

After a timeout of 1 second if the image doesn't have at least 100px lines it is considered an incomplete image and not drawn. This is because sometimes a transient byte is detected on the serial line and this shouldn't be drawn.

Also note that the image is drawn horizontally mirrored for some reason.